hedgedoc: fix config
Having the setting `dhParamPath = null` present in the config results in
having this line in Hedgedoc's config.json file.
If Hedgedoc recognizes that a certain path-related setting (like dhParamPath)
it tries to load the file.

HOWEVER if the path is not a string but `null` the whole application
crashes and leaves the administrator with nothing but a vacuous error
message:
  The "paths[1]" argument must be of type string. Received null

There is one exception to the above: the setting parameter "path"
specifying the socket path (in case one does not want to use IP network
communication).
This path is actually allowed to be of value `null`.

[Utz: changed title]